requirements.txt
numpy
openai
pyaudio
pyttsx3
python-dotenv 